Fixed crash when reading/inserting routines, needs cleanup still
Appears to have been an issue with the DB version. A faulty onUpgrade() method in GymDbHelper.java was causing the app to not correctly recreate the updated DB. Fixed by reinstalling on my phone, but not a perfect fix obviously. Probably will add a "Real Later" issue for this.

NEED TO CLEANUP ALL THE TEST CODE. I hard coded some strings (like category; for some reason that was coming through as null) and used .execSQL() instead of inserting via ContentValues. Gotta clean all that up later