Adding functionality to add exercise to a routine
Added the options menu to a routine detail. There's an option to "Add an exercise to this routine."

This will have to be done by passing a boolean extra (representing if we came from "Add an exercise to routine") and an integer extra (representing the DB id of the routine to add to) with the intent to the CategoryListActivity and then through to the list of exercises.

If the intent extra (called "addingToRoutine") is true, instead of launching the exercise detail activity, insert a record into the exercises_in_routines table with the key value pair of the exercise's _id field (use getDbId()) and the routine's _id field.

NOTE: Less than half of this is done so far; this is just the design. Will finish implementing on next commit.